Renamed python package from google.cloud to opentelemetry
